[feat]: YOLO 화재관련 재라밸링 기능, 유틸 추가
