+++ feat_detect_fire/data.yaml
@@ -1,0 +1,22 @@
+# data.yaml
+
+# 화재연기 / 일반연기 / 무관객체 재라벨링 클래스
+
+# 데이터셋 경로
+train: data/train/images
+test: data/test/images
+val: data/val/images
+
+# Classes
+names:
+  0: fire
+  1: fire_smoke
+  2: general_smoke
+  3: irrelevant
+# names : ['fire', 'fire_smoke', 'general_smoke', 'irrelevant']
+
+# Counts
+nc: 4 # number of classes
+train_count: 572
+test_count: 78
+val_count: 100
